Conflicts:
	project/target/streams/$global/ivyConfiguration/$global/streams/out
	project/target/streams/$global/update/$global/streams/out
	project/target/streams/compile/dependencyClasspath/$global/streams/export
	project/target/streams/compile/externalDependencyClasspath/$global/streams/export
	project/target/streams/compile/managedClasspath/$global/streams/export

+++ src/main/scala/classifier/adaGradTrainer.scala
@@ -20,7 +20,7 @@ object adaGradTrainer {
     }
   }
 
-  def train(documents:Seq[Document], eta:Double, lambda:Double, iterMax:Int, featureCount:Int, featureExtractor: (Document, Int, Int) => Seq[Double], lossFunction:(Document, Int, Int) => Double):Seq[Double] = {
+  def train(documents:Seq[Document], eta:Double, lambda:Double, iterMax:Int, featureCount:Int, featureExtractor: (Document, Int, Int) => Seq[Int], lossFunction:(Document, Int, Int) => Double):Seq[Double] = {
     var grad =  List.fill(featureCount)(0.0)
     var dgrad = List.fill(featureCount)(0.0)
     var weights =     List.fill(featureCount)(0.0)
@@ -44,7 +44,7 @@ object adaGradTrainer {
     weights.toList
   }
 
-  def computeGradient(document:Document,gradient:List[Double],lossFunction:(Document, Int, Int) => Double, scores:Seq[Seq[Double]], featureExtractor: (Document, Int, Int) => Seq[Double]):List[Double] = {
+  def computeGradient(document:Document,gradient:List[Double],lossFunction:(Document, Int, Int) => Double, scores:Seq[Seq[Double]], featureExtractor: (Document, Int, Int) => Seq[Int]):List[Double] = {
   /*
     A gradient is marginals Gold - marginals predicted. Thus the max is 0 when there is no error amount. If the score is high and the loss is high, the error wil be grown
    */
@@ -87,7 +87,7 @@ object adaGradTrainer {
         unregMij}}.map(unregularized=> unregularized/marginalOveri)}}
   }
 
-  def updateGradient(gradient:List[Double], features:Seq[Double], weight:Double): List[Double] = {
+  def updateGradient(gradient:List[Double], features:Seq[Int], weight:Double): List[Double] = {
     gradient.zip(features).map{case ((gk:Double,fk:Double)) => if (fk >= 1.0e-20) gk+weight else gk}
   }
 }
